imapserver: move CAPABILITY, IDLE and LOGIN handlers to separate files
