+++ imapserver/idle.go
@@ -1,0 +1,32 @@
+package imapserver
+
+import (
+	"github.com/emersion/go-imap/v2"
+	"github.com/emersion/go-imap/v2/internal/imapwire"
+)
+
+func (c *conn) handleIdle(dec *imapwire.Decoder) error {
+	if !dec.ExpectCRLF() {
+		return dec.Err()
+	}
+
+	if err := c.checkState(imap.ConnStateAuthenticated); err != nil {
+		return err
+	}
+
+	enc := newResponseEncoder(c)
+	defer enc.end()
+
+	if err := writeContReq(enc.Encoder, "idling"); err != nil {
+		return err
+	}
+
+	line, isPrefix, err := c.br.ReadLine()
+	if err != nil {
+		return err
+	} else if isPrefix || string(line) != "DONE" {
+		return newClientBugError("Syntax error: expected DONE to end IDLE command")
+	}
+
+	return nil
+}
